refactor: add infrastructure adapters
