+++ src/infrastructure/logging/PinoLoggerFactory.ts
@@ -4,9 +4,9 @@ import pino, { type LevelWithSilent, type Logger as Pino } from 'pino';
 import {
   ENV_SERVICE_ID,
   type EnvService,
-} from '../../interfaces/env/EnvService.interface';
-import type { Logger } from '../../interfaces/logging/Logger.interface';
-import { type LoggerFactory } from '../../interfaces/logging/LoggerFactory.interface';
+} from '../../application/interfaces/env/EnvService.interface';
+import type { Logger } from '../../application/interfaces/logging/Logger.interface';
+import { type LoggerFactory } from '../../application/interfaces/logging/LoggerFactory.interface';
 import { PinoLogger } from './PinoLogger';
 
 @injectable()
