+++ src/infrastructure/repositories/SQLiteChatRepository.ts
@@ -3,7 +3,7 @@ import type { Database } from 'sqlite';
 
 import type { ChatEntity } from '../../domain/entities/ChatEntity';
 import type { ChatRepository } from '../../domain/repositories/ChatRepository.interface';
-import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
+import { DB_PROVIDER_ID, type SQLiteDbProvider } from './DbProvider';
 
 @injectable()
 export class SQLiteChatRepository implements ChatRepository {
