+++ src/infrastructure/repositories/SQLiteMessageRepository.ts
@@ -4,7 +4,7 @@ import type { Database } from 'sqlite';
 import type { ChatMessage } from '../../application/interfaces/ai/AIService.interface';
 import type { StoredMessage } from '../../application/interfaces/messages/StoredMessage.interface';
 import type { MessageRepository } from '../../domain/repositories/MessageRepository.interface';
-import { DB_PROVIDER_ID, type SQLiteDbProvider } from '../DbProvider';
+import { DB_PROVIDER_ID, type SQLiteDbProvider } from './DbProvider';
 
 @injectable()
 export class SQLiteMessageRepository implements MessageRepository {
